Implemented the indexing of each book individually. Now I need to perform experiments to see what scores I can get and how I can improve them
