+++ general/general/books.py
@@ -1,7 +1,10 @@
 import requests
 from pathlib import Path
-from loguru import logger
 
+from loguru import logger
+from dropbox import Dropbox
+
+from general.config import general_config
 from general.paths import set_paths, make_data_directories
 
 
@@ -17,8 +20,7 @@ class Book:
     def __get_file_path__(self):
         return set_paths(from_scratch=False, general=True)["raw_data"]
 
-    def download(self):
-        from general.remote_storage import upload_to_dropbox, download_from_dropbox 
+    def download(self, upload: bool):
         
         if Path(self.file_path).exists():
             logger.success(f'"{self.title}" is already saved to disk')
@@ -40,7 +42,8 @@ class Book:
                 logger.warning(f'Attempting to download "{self.title}" from Dropbox')
                 download_from_dropbox(book=self) 
        
-        upload_to_dropbox(book=self)
+        if upload:
+            upload_to_dropbox(book=self)
 
     def __find_non_core_pages__(self) -> tuple[int, int]: 
         
@@ -52,6 +55,36 @@ class Book:
         
         assert self.file_name in book_and_non_core_pages.keys()
         return book_and_non_core_pages[self.file_name]
+
+
+box = Dropbox(general_config.dropbox_access_token)
+
+def upload_to_dropbox(book: Book) -> None:
+
+    remote_file_path = f"/{book.file_name}.pdf"
+
+    try:
+        logger.info(f'Checking whether {book.title} has already been uploaded')
+        metadata = box.files_get_metadata(path=remote_file_path) 
+
+        if remote_file_path in metadata.values():
+            logger.success(f'"{book.title}" is already on Dropbox')
+
+    except Exception as error:
+        logger.error(error)
+        logger.info(f'File not found. Attempting to upload "{book.title}" to Dropbox')
+
+        with open(book.file_path, "rb") as file:
+           box.files_upload(f=file.read(), path=remote_file_path)
+
+        logger.success(f"'{book.title}'uploaded to Dropbox")
+
+
+def download_from_dropbox(book: Book) -> None:
+    metadata, response = box.files_download(f"/{book.file_name}.pdf")
+
+    with open(book.file_path, "wb") as file:
+        file.write(response.content)
 
    
 neo_colonialism = Book(
@@ -76,5 +109,5 @@ africa_unite = Book(
 if __name__ == "__main__":
     make_data_directories(from_scratch=False, general=True)
     for book in [neo_colonialism, dark_days, africa_unite]:
-        book.download()
+        book.download(upload=True)
